git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/branches/geant4/_symbols/09-04-ref-09_branch@52909 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/visualization/OpenGL/include/G4OpenGLQtViewer.hh
@@ -99,7 +99,7 @@ public:
   void saveVideo();
   bool generateMpegEncoderParameters();
   void displayRecordingStatus();
-  void DrawText(const char * ,int x,int y,int z, int size);
+  void DrawText(const char * ,double x,double y,double z, double size);
 public:
   void G4MousePressEvent(QMouseEvent *event);
   void G4wheelEvent (QWheelEvent * event);
