+++ source/visualization/OpenGL/src/G4OpenGLXViewer.cc
@@ -342,12 +342,12 @@ void G4OpenGLXViewer::CreateFontLists () {
 }
 
 
-void G4OpenGLXViewer::DrawText(const char * textString,int x,int y,int z, int size){
+void G4OpenGLXViewer::DrawText(const char * textString,double x,double y,double z, double size){
   
   // gl2ps or GL window ?
   if (! drawGl2psText(textString,size)) {
 
-    G4int font_base = G4OpenGLFontBaseStore::GetFontBase(this,size);
+    G4int font_base = G4OpenGLFontBaseStore::GetFontBase(this,(int)size);
     if (font_base < 0) {
       static G4int callCount = 0;
       ++callCount;
